Start work on some dynastat controls
Build the dynastat controls as a seperate in-repo-addon so we could export it in the future.

Add a basic webrtc-controls service to deal with the connection.

Create a sensor display component for rendering sensor values.
This is yet to be tested in the real world, currently it just displays the coloured dots in the correct places and is almost a direct port from the .coffee file.